Move findKeywordInFirstParagraph to match folder
and adjust its reference in other files
